+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Ease Animation curves preset
 - Animation Curve Extensions
+- Position (Shake) Animation component
 - Position Animation component
 - Scale Animation component
 - TextMesh Pro Visible Characters Animation component